+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ ] add support for `typing.IO` objects instead of too narrow file path
 - [ ] move multithreading logic from Python to Rust
 - [x] cover with unit-tests
+- [x] feature: add special `-` argument for reading `file_path` list from STDIN for CLI application
 
 ## [0.1.0] — 2025-04-16